Thrown together some property wrappers for rigid bodies and shapes for display on a property grid.
Forced selection on the scene graph view whenever the tree view updates, to signal to the rest of the application.
Hooked up selection with the mouse and from the tree view to drive the property grid view on the UI.
Still to do - proper property grid display helpers for all of the types that compose rigid bodies/shapes etc. as the collections are appearing as editable things rather than an expandable list of sub members. Collections are also appearing as an editable collection, which opens a new dialog, rather than again, displaying the list of child elements as an expandable property.