Initial submit of a C++ with SDL, ImGui and DX11 application framework
Initial submit of Odin - the C++ version of the debugger. It uses the ImGui docker branch for window management, to allow for Visual Studio style window docking and managment. SDL is used to create the main application window and DX11 for rendering. Made using the ImGui SDL + DX11 sample. Only changes so far are to remove the ImGui window and enable dockspace at the root of the application so that everything can dock everywhere.